seq-count and list-from-roots ship, and `count` stops being O(n) on the ports
Row 5 of the port order, bar `seq` itself, which still waits on hole 5.

THE PORTS HAD NO FAST PATH. Rust's `seq_count` asked `count_hint` first and
returned without walking when a collection knew its own size; the JVM and CLR
went straight to the walk. Same function, same answers, a different complexity
class -- `(count v)` on a vector was O(1) on native and O(n) on both ports.
Nothing could have caught it: conform diffs ANSWERS, and both answers are
right. Generating the function from one source is what removes it, which is
the argument this whole port rests on, showing up as a measurement instead of
a claim.

`list_from_roots` needed a convergence. Rust wrote it `for i in (0..n).rev()`
and both ports `for (i = n - 1; i >= 0; i--)`; kin's `for` is half-open and
ascending only. The loop now counts up and the INDEX counts down, which is
cheaper than a descending loop form and reproduces the same order of conses.
Reversing it in the source flips the probe's `lfr=65,66,67` to `67,66,65`.

`n` is tagged `I32` and not `RootIx`. It is a COUNT; `RootIx` means an index
into the shadow stack, and borrowing it would have compiled on all three and
told a reader something false -- the reason `Ty` and `Cat` are two tags. That
costs Rust's signature `n: u32` and four call sites an `as u32`, and it sets
the shape for the three `*_from_roots` still to come in rows 6 and 8.

WHAT THE DRIVERS DID NOT DO AT FIRST. Both probes compiled the new functions
without calling them, which is a gate that cannot fail. They call them now,
and the counted chain LIES: it is two cells long and says nine, so `cnt_hint=9`
proves the hint was read where a truthful count would answer 2 down either
path and could not tell the fast path from its absence.

Also fixed in kin (8808953): a parameter named `base` emitted `@base` in every
C# body and `int base` in the signature, so the CLR would not parse. All three
signature builders go through the target's `:local-name` now.

Gates: bin/test 0, conform 26 rows + 13 GC-stress, 0 FAIL, language suite
84/84 matched by both ports, check-kin green, coverage 217/256 cold against
the 217/256 baseline.

+++ runtime/src/codec.rs
@@ -499,7 +499,7 @@ impl Rt {
                 }
                 let out = match tag {
                     K_VECTOR => self.vec_from_roots(base, n),
-                    K_LIST => self.list_from_roots(base, n),
+                    K_LIST => self.list_from_roots(base, n as u32),
                     _ => {
                         // No `set_from_roots`, so it is built by conj -- and
                         // the accumulator lives on the ROOT STACK, because
